refactor: extract custom hooks out of the main component file
